Move spill related functions to spill.rs
